# specs/005-cloud-event-deployment/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Chatbot API - Phase V Extensions
  version: 5.0.0
  description: Advanced task features, recurring tasks, tags, search, and event-driven endpoints
paths:
  /api/{user_id}/tasks:
    get:
      summary: List tasks with filters and sorting
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
        - name: tag
          in: query
          schema:
            type: integer
          description: Tag ID to filter by
        - name: overdue
          in: query
          schema:
            type: boolean
        - name: due_soon
          in: query
          schema:
            type: boolean
          description: Due within 24 hours
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, title, due_date, priority, completed]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      summary: Create task with priority, tags, due date
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /api/{user_id}/tasks/{task_id}:
    put:
      summary: Update task including priority, tags, due date
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...
  /api/{user_id}/tags:
    get:
      summary: List user tags
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'

    post:
      summary: Create tag
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
  /api/{user_id}/tags/{tag_id}:
    delete:
      summary: Delete tag
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: tag_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Tag deleted
  /api/{user_id}/recurring-tasks:
    get:
      summary: List recurring task templates
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of recurring tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringTask'

    post:
      summary: Create recurring task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringTaskCreate'
      responses:
        '201':
          description: Recurring task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringTask'
  /api/{user_id}/recurring-tasks/{id}:
    put:
      summary: Update recurring task pattern
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringTaskUpdate'
      responses:
        '200':
          description: Recurring task updated

    delete:
      summary: Stop recurring task (keep existing instances)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Recurring task stopped
...
  /events/task-events:
    post:
      summary: Dapr event handler for task events
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskEvent'
      responses:
        '200':
          description: Event processed
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [high, medium, low]
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        due_time:
          type: string
          nullable: true
        recurring_task_id:
          type: integer
          nullable: true
        is_recurring_instance:
          type: boolean
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        created_at:
          type: string
          format: date-time
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        due_date:
          type: string
          format: date
        due_time:
          type: string
        tag_ids:
          type: array
          items:
            type: integer
        recurring:
          type: boolean
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
        recurrence_interval:
          type: integer
        days_of_week:
          type: array
          items:
            type: integer
        day_of_month:
          type: integer
        end_date:
          type: string
          format: date
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        completed:
          type: boolean
        due_date:
          type: string
          format: date
        due_time:
          type: string
        tag_ids:
          type: array
          items:
            type: integer
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
          description: Hex color (#RRGGBB)
        created_at:
          type: string
          format: date-time
    TagCreate:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
    RecurringTask:
      type: object
      properties:
        id:
          type: integer
        task_template:
          type: object
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
        interval:
          type: integer
        days_of_week:
          type: array
          items:
            type: integer
        day_of_month:
          type: integer
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    RecurringTaskCreate:
      type: object
      required: [task_template, recurrence_pattern]
      properties:
        task_template:
          type: object
          required: [title]
          properties:
            title:
              type: string
            description:
              type: string
            priority:
              type: string
            tag_ids:
              type: array
              items:
                type: integer
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
        interval:
          type: integer
          minimum: 1
          default: 1
        days_of_week:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
        day_of_month:
          type: integer
          minimum: 1
          maximum: 31
        end_date:
          type: string
          format: date
    RecurringTaskUpdate:
      type: object
      properties:
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
        interval:
          type: integer
          minimum: 1
        days_of_week:
          type: array
          items:
            type: integer
        day_of_month:
          type: integer
        end_date:
          type: string
          format: date
        active:
          type: boolean
...
    TaskEvent:
      type: object
      properties:
        schema_version:
          type: string
        event_type:
          type: string
          enum: [created, updated, completed, deleted]
        task_id:
          type: integer
        user_id:
          type: string
        task_data:
          type: object
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          properties:
            source:
              type: string
            correlation_id:
              type: string
              format: uuid